# docs/swagger.yaml
basePath: /
definitions:
...
  models.Category:
    properties:
      id:
        type: string
      image:
        type: string
      name:
        type: string
      parentId:
        type: string
      slug:
        type: string
    type: object
...
  models.Manufacturer:
    properties:
      id:
        type: string
      logo:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  models.Product:
    properties:
      availability:
        type: string
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: string
      description:
        type: string
      features:
        items:
          type: string
        type: array
      id:
        type: string
      image:
        items:
          type: string
        type: array
      manufacturer:
        $ref: '#/definitions/models.Manufacturer'
      manufacturerId:
        type: string
      name:
        type: string
      oldPrice:
        type: integer
      price:
        type: integer
      rating:
        type: number
      reviews:
        type: integer
      sku:
        type: string
      slug:
        type: string
      stock:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Noble Group Services e-commerce platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Noble Group Services API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Get a list of products with optional filtering
      parameters:
      - description: Category Slug
        in: query
        name: category
        type: string
      - description: Manufacturer Slug
        in: query
        name: manufacturer
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Only in stock
        in: query
        name: inStockOnly
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Get list of products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Create a product
      tags:
      - products
  /products/{id}:
    delete:
      description: Delete a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Product not found
          schema:
            type: string
      summary: Delete product
      tags:
      - products
    get:
      description: Get details of a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "404":
          description: Product not found
          schema:
            type: string
      summary: Get product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Product not found
          schema:
            type: string
      summary: Update product
      tags:
      - products
...
swagger: "2.0"